add base block test suite
